+++ CHANGELOG.md
@@ -6,7 +6,17 @@ Add new items at the end of the relevant section under **Unreleased**.
 
 ## [Unreleased]
 
-*No new changes.*
+### Additions
+
+*No additions*
+
+### Changes
+
+- Support extraction of new "Summary of changes" section ([#79])
+
+### Fixes
+
+*No fixes*
 
 ---
 
@@ -71,3 +81,4 @@ This changelog's format is based on [Keep a Changelog](https://keepachangelog.co
 [#70]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/70
 [#72]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/72
 [#74]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/74
+[#79]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/79